requirements.txt
requests
scraperapi-sdk
beautifulsoup4
setuptools